make component to join github and rigobot
